renaming MapModal comp and adding error message to modal
