--- react/src/components/MapModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from './MapModal';
