+++ react/src/pages/MainMenu/MainMenu.test.tsx
@@ -5,13 +5,16 @@ import { QueryClientProvider } from 'react-query';
 import { testQueryClient } from '../../testUtil';
 import { Provider } from 'react-redux';
 import store from '../../redux/store';
+import { BrowserRouter as Router } from 'react-router-dom';
 
 test('renders menu', () => {
   const { getByText } = render(
     <Provider store={store}>
-      <QueryClientProvider client={testQueryClient}>
-        <MainMenu />
-      </QueryClientProvider>
+      <Router>
+        <QueryClientProvider client={testQueryClient}>
+          <MainMenu />
+        </QueryClientProvider>
+      </Router>
     </Provider>
   );
   expect(getByText(/Main Menu/)).toBeDefined();
