+++ react/src/pages/MainMenu/MainMenu.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useRef } from 'react';
 import {
   LoadingSpinner,
   InlineMessage,
@@ -6,7 +6,9 @@ import {
   Button,
 } from '../../core-components';
 import { useProjects } from '../../hooks';
-import MapModal from '../../components/MapModal';
+import { useNavigate } from 'react-router-dom';
+import { Project } from '../../types';
+import CreateMapModal from '../../components/CreateMapModal/CreateMapModal';
 import useCreateProject from '../../hooks/projects/useCreateProject';
 
 function MainMenu() {
@@ -14,19 +16,37 @@ function MainMenu() {
   const { mutate: createProject, isLoading: isCreatingProject } =
     useCreateProject();
   const [isModalOpen, setIsModalOpen] = useState(false);
+  const createMapModalRef = useRef<{
+    setErrorMessage: (message: string) => void;
+  }>(null);
 
-  const toggleModal = () => setIsModalOpen(!isModalOpen);
+  const toggleModal = () => {
+    if (createMapModalRef.current) {
+      createMapModalRef.current.setErrorMessage('');
+    }
+    setIsModalOpen(!isModalOpen);
+  };
+
+  const navigate = useNavigate();
 
   const handleCreateProject = (projectData) => {
     createProject(projectData, {
-      onSuccess: () => {
+      onSuccess: (newProject: Project) => {
         // TO-ADD: Handle success (e.g., show success message, refetch projects, etc.)
-        console.log('Data after submit', projectData);
-        toggleModal();
+        navigate(`/project/${newProject.uuid}`);
       },
       onError: (err) => {
         // Handle error (e.g., show error message)
         console.error('Error creating project:', err);
+        let errorMessage = 'An error occurred while creating the project.';
+        if (err?.response?.status === 409) {
+          errorMessage = 'That folder is already syncing with a different map.';
+        } else if (err?.response?.status === 500) {
+          errorMessage = 'Internal server error. Please contact support.';
+        } else {
+          errorMessage = 'Something went wrong. Please contact support.';
+        }
+        createMapModalRef.current?.setErrorMessage(errorMessage);
       },
     });
   };
@@ -58,7 +78,8 @@ function MainMenu() {
       >
         Create Map
       </Button>
-      <MapModal
+      <CreateMapModal
+        ref={createMapModalRef}
         isOpen={isModalOpen}
         toggle={toggleModal}
         onSubmit={handleCreateProject}
